services/coordination/pyproject.toml
[project]
name = "coordination"
version = "0.1.0"
description = "Coordination microservice for IoT Edge Gateway — cluster consensus, leader election, gossip membership, config sync, OTA"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "shared",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
include = ["coordination*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
coordination = ["py.typed"]

# Local path dependency for development installs
# Use: pip install -e ../../shared && pip install -e .
[tool.pip.find-links]
shared = "../../shared"
